fixing mapping function

+++ src/org/erhsroboticsclub/robo2013/utilities/MathX.java
@@ -78,7 +78,7 @@ public class MathX {
         return d2;
     }
 
-    public static double map(double x, double in_min, long in_max, long out_min, long out_max) {
+    public static double map(double x, double in_min, double in_max, double out_min, double out_max) {
         return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
     }
